tests/wpt/web-platform-tests/cssom-view/historical.html

<!doctype html>
<meta charset="utf-8">
<title>Historical features</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
test(function() {
  assert_false("DOMRectList" in self);
}, "Support for DOMRectList");
</script>
